+++ src/main/java/se/terrassorkestern/notgen2/song/SongController.java
@@ -5,6 +5,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -37,10 +40,13 @@ public class SongController {
   }
 
   @GetMapping("/delete")
-  public String songDelete(@RequestParam("id") Integer id, Model model) {
-    Song song = songRepository.findById(id).get();
-    log.info("Tar bort låt " + song.getTitle() + " [" + song.getId() + "]");
-    songRepository.delete(song);
+  public String songDelete(@RequestParam("id") Integer id, Model model,
+      @AuthenticationPrincipal UserDetails userDetails) {
+    if (userDetails.getAuthorities().contains(new SimpleGrantedAuthority("EDIT_SONG"))) {
+      Song song = songRepository.findById(id).get();
+      log.info("Tar bort låt " + song.getTitle() + " [" + song.getId() + "]");
+      songRepository.delete(song);
+    }
     return "redirect:/song/list";
   }
 
@@ -64,20 +70,23 @@ public class SongController {
   }
 
   @PostMapping("/save")
-  public String songSave(@Valid @ModelAttribute Song song, Errors errors, Model model) {
+  public String songSave(@Valid @ModelAttribute Song song, Errors errors, Model model,
+      @AuthenticationPrincipal UserDetails userDetails) {
     if (errors.hasErrors()) {
       model.addAttribute("allInstruments", instrumentRepository.findByOrderByStandardDescSortOrder());
       return "songEdit";
     }
-    log.info("Sparar låt " + song.getTitle() + " [" + song.getId() + "]");
-    // scorPart måste fixas till efter formuläret
-    for (ScorePart scorePart : song.getScoreParts()) {
-      Instrument instrument = scorePart.getInstrument();
-      scorePart.setId(new ScorePartId(song.getId(), instrument.getId()));
-      scorePart.setSong(song);
-      scorePart.setInstrument(instrumentRepository.findById(instrument.getId()).get());
+    if (userDetails.getAuthorities().contains(new SimpleGrantedAuthority("EDIT_SONG"))) {
+      log.info("Sparar låt " + song.getTitle() + " [" + song.getId() + "]");
+      // scorPart måste fixas till efter formuläret
+      for (ScorePart scorePart : song.getScoreParts()) {
+        Instrument instrument = scorePart.getInstrument();
+        scorePart.setId(new ScorePartId(song.getId(), instrument.getId()));
+        scorePart.setSong(song);
+        scorePart.setInstrument(instrumentRepository.findById(instrument.getId()).get());
+      }
+      songRepository.save(song);
     }
-    songRepository.save(song);
     return "redirect:/song/list";
   }
